Implementing WAIT, completing verb refactoring, and starting barriers
It is possible to indicate an actual duration, which will be parsed as
a PostgreSQL interval, as in "wait 13 hours". Amusingly, this allows
adventures spanning a long time, where the age of characters can be
tracked and possibly be relevant.

Barriers are still WIP; they are a prerequisite for OPEN and CLOSE. We
provide an auto_close attribute which reflects the way most doors work
nowadays.

Verb refactoring has allowed abbreviations, e.g. "wa" for WAIT. The
only priority is given to GO directions, i.e. "w" means WEST even if
it could be seen as an abbreviation of WAIT. It is an error to use an
abbreviation that selects more than one verb.

Verb refactoring will also allow translating the adventure to
languages other than english.

Signed-off-by: Gianni Ciolli <gc@inatna.it>