Eliminating a type, introducing action history, and refactoring init
We no longer need a type declaration as the action history already
defines the right columns for the action data type.

Now the initialisation is wrapped in a procedure, with a generic name
pgif_init(), so it can be repeated.

Moving towards the separation between static definition and mutable
state.

Signed-off-by: Gianni Ciolli <gc@inatna.it>