Refactoring parser match for transitive verbs
Now we use a single match() function for matching the object for each
transitive verb, and for matching the verb itself.

This simplifies error messages, and also allows abbreviations, as well
as more intelligent defaults such as not requiring the specification
of the object if there is only one possible candidate.

Signed-off-by: Gianni Ciolli <gc@inatna.it>