The state of a game can now be dumped and restored
We have appropriately marked all the sequences and tables that contain
state; hence an ongoing game can just be saved with pg_dump, and
reloaded accordingly.

Before loading a dump, target tables must be truncated. This can be
done either by dropping and recreating the extensions, or by issuing
appropriate SQL such as

    TRUNCATE actions, object_state, barrier_state;

Signed-off-by: Gianni Ciolli <gc@inatna.it>